requirements.txt
# ── VM (notify.py) ────────────────────────────────────────────────────────────
# Install on Oracle Cloud with: pip install -r requirements.txt --no-deps
# These are the only packages needed to run the scraper/notifier daemon.
httpx==0.27.0
requests==2.32.3
beautifulsoup4==4.12.3
feedparser==6.0.11
PyYAML==6.0.2

# ── Local dashboard (run.py) ──────────────────────────────────────────────────
# Required only on your local machine.
flask==3.1.3
flask-sqlalchemy==3.1.1
telethon==1.37.0
apscheduler==3.10.4
sqlalchemy==2.0.49
